Modularized modal and settings
Moved PdfToImageModal and PluginSettingPage classes from main.ts into separate modal.ts and settings.ts files for better code organization.
