+++ main.ts
@@ -1,28 +1,10 @@
-import { App, Editor, MarkdownView, Modal, Notice, Plugin, TFile, normalizePath, moment, PluginSettingTab, Setting, setIcon, FuzzySuggestModal, TFolder, loadPdfJs, FileManager } from 'obsidian';
-
-interface PluginSettings {
-	enableHeaders: boolean;
-	headerSize: string;
-	headerExtractionSensitive: number;
-	removeHeaderDuplicates: boolean;
-	imageResolution: number;
-	afterImage: number;
-	insertionMethod: string;
-}
-
-const DEFAULT_SETTINGS: PluginSettings = {
-	enableHeaders: false,
-	headerSize: "#",
-	headerExtractionSensitive: 1.2,
-	removeHeaderDuplicates: false,
-	imageResolution: 1,
-	afterImage: 0,
-	insertionMethod: 'Procedual'
-}
+import { App, Editor, MarkdownView, Notice, Plugin, normalizePath, loadPdfJs, FileManager } from 'obsidian';
+import { PluginSettings, DEFAULT_SETTINGS, PluginSettingPage } from './settings';
+import { PdfToImageModal } from './modal';
 
 export default class Pdf2Image extends Plugin {
 	settings: PluginSettings;
-	private ribbonEl: HTMLElement | null = null;
+	ribbonEl: HTMLElement | null = null;
 	pdfjsLib: any;
 	fileManager: FileManager;
 
@@ -295,292 +277,3 @@ export default class Pdf2Image extends Plugin {
 }
 
 
-/**
- * A modal dialog for selecting a PDF file and converting it to images.
- * 
- * This modal allows the user to select a PDF file from their file system and submit it for conversion.
- * It provides a file input element for selecting the PDF and a button to trigger the conversion process.
- * 
- * @extends Modal
- */
-class PdfToImageModal extends Modal {
-	private file: File | null = null;
-	private imageQuality: number;
-
-	constructor(app: App, private onSubmit: (file: File, imageQuality: number) => void, defaultImageQuality: number) {
-		super(app);
-		this.imageQuality = defaultImageQuality;
-	}
-
-	/**
-	 * Handles the opening of the plugin interface.
-	 * 
-	 * This method sets up the UI elements for the user to select a PDF file and initiate the conversion process.
-	 * It creates a header, a file input for PDF selection, and a submit button.
-	 * 
-	 * - The file input allows the user to select a PDF file.
-	 * - The submit button triggers the conversion process if a file is selected, otherwise it shows a notice.
-	 * 
-	 * @remarks
-	 * - The selected file is stored in `this.file`.
-	 * - The `onSubmit` method is called with the selected file when the submit button is clicked.
-	 * - If no file is selected, a notice is displayed to the user.
-	 */
-	onOpen() {
-		const { contentEl } = this;
-		const header = contentEl.createEl('h2', { text: 'Select a PDF file to convert' });
-		header.style.textAlign = 'center';
-		header.style.marginTop = '0px';
-
-		// File input section
-		const fileSection = contentEl.createDiv();
-		// fileSection.createEl('label', { text: 'PDF File: ' });
-
-		const fileInputWrapper = fileSection.createDiv();
-		fileInputWrapper.style.marginTop = '5px';
-		fileInputWrapper.style.textAlign = 'center';
-
-		// Hide the actual file input
-		const fileInput = fileInputWrapper.createEl('input', { 
-			type: 'file', 
-			attr: { accept: '.pdf' } 
-		});
-		fileInput.style.display = 'none';
-
-		// Create a custom button
-        const customFileButton = fileInputWrapper.createEl('button', { 
-            text: 'Choose PDF File',
-            attr: {
-                'aria-label': 'Choose PDF file to upload',
-                'aria-controls': fileInput.id || 'pdf-file-input'
-            }
-        });
-		customFileButton.style.cssText = `
-            padding: 8px 16px;
-            cursor: pointer;
-            font-size: 14px;
-        `;
-
-		fileInput.id = fileInput.id || 'pdf-file-input';
-
-		customFileButton.onclick = (e) => {
-			e.preventDefault();
-			fileInput.click();
-		};
-
-		// Create custom file name display
-		const fileNameDisplay = fileInputWrapper.createDiv();
-		fileNameDisplay.style.cssText = `
-			margin-top: 8px;
-			font-size: 0.9em;
-			color: #666;
-			font-style: italic;
-		`;
-		fileNameDisplay.textContent = 'No file chosen';
-
-		fileInput.onchange = () => {
-			if (fileInput.files && fileInput.files.length > 0) {
-				this.file = fileInput.files[0];
-				fileNameDisplay.textContent = `${fileInput.files[0].name}`;
-				fileNameDisplay.style.color = '#7a7a7a';
-				fileNameDisplay.style.fontStyle = 'normal';
-			} else {
-				fileNameDisplay.textContent = 'No file chosen';
-				fileNameDisplay.style.color = '#666';
-				fileNameDisplay.style.fontStyle = 'italic';
-			}
-		};
-
-		// Image quality dropdown section
-		const qualitySection = contentEl.createDiv();
-		qualitySection.style.marginTop = '15px';
-		qualitySection.style.textAlign = 'center';
-		qualitySection.createEl('label', { text: 'Image Quality' });
-		qualitySection.createEl('br');
-		const qualitySelect = qualitySection.createEl('select');
-		qualitySelect.style.marginTop = '5px';
-		qualitySelect.style.padding = '5px';
-		qualitySelect.style.cursor = 'pointer';
-		
-		// Add options to the dropdown
-		const qualityOptions = [
-			{ value: '0.5', text: '0.5x' },
-			{ value: '0.75', text: '0.75x' },
-			{ value: '1', text: '1x' },
-			{ value: '1.5', text: '1.5x' },
-			{ value: '2', text: '2x' }
-		];
-
-		qualityOptions.forEach(option => {
-			const optionEl = qualitySelect.createEl('option', { 
-				value: option.value, 
-				text: option.text 
-			});
-			optionEl.style.textAlign = 'center';
-			if (parseFloat(option.value) === this.imageQuality) {
-				optionEl.selected = true;
-			}
-		});
-
-		qualitySelect.onchange = () => {
-			this.imageQuality = parseFloat(qualitySelect.value);
-		};
-
-		// Add description for image quality
-		// const qualityDesc = qualitySection.createEl('div');
-		// qualityDesc.style.fontSize = '0.8em';
-		// qualityDesc.style.color = '#888';
-		// qualityDesc.style.marginTop = '5px';
-		// qualityDesc.textContent = 'Lower = faster and smaller file size, higher = slower and bigger file size';
-
-		// Submit button
-		const buttonSection = contentEl.createDiv();
-		buttonSection.style.marginTop = '20px';
-		buttonSection.style.textAlign = 'center';
-		const submitButton = buttonSection.createEl('button', { text: 'Convert' });
-		submitButton.style.padding = '10px 20px';
-		submitButton.style.cursor = 'pointer';
-		submitButton.onclick = () => {
-			if (this.file) {
-				this.onSubmit(this.file, this.imageQuality);
-				this.close();
-			} else {
-				new Notice('Please select a PDF file');
-			}
-		};
-	}
-
-	/**
-	 * Handles the closing of the plugin's UI component.
-	 * This method is called when the component is closed and is responsible for cleaning up the content element.
-	 * It empties the content element to ensure no residual elements remain.
-	 */
-	onClose() {
-		const { contentEl } = this;
-		contentEl.empty();
-	}
-}
-
-/**
- * Represents the settings page for the plugin.
- *
- * @class PluginSettingPage
- * @extends {PluginSettingTab}
- */
-class PluginSettingPage extends PluginSettingTab {
-	plugin: Pdf2Image;
-
-	constructor(app: App, plugin: Pdf2Image) {
-		super(app, plugin);
-		this.plugin = plugin;
-	}
-
-	display(): void {
-		const { containerEl } = this;
-
-		containerEl.empty();
-
-		// Image Quality setting
-		new Setting(containerEl)
-			.setName('Image quality')
-			.setDesc('The quality of the images to be generated. Lower = faster and smaller file size, higher = slower and bigger file size. The default is 1x.')
-			.addDropdown(dropdown => dropdown
-				.addOption('0.5', '0.5x')
-				.addOption('0.75', '0.75x')
-				.addOption('1', '1x')
-				.addOption('1.5', '1.5x')
-				.addOption('2', '2x')
-				.setValue(this.plugin.settings.imageResolution.toString())
-				.onChange(async (value) => {
-					this.plugin.settings.imageResolution = parseFloat(value);
-					await this.plugin.saveSettings();
-				}));
-
-		// Insertion Method setting
-		new Setting(containerEl)
-			.setName('Image insertion method')
-			.setDesc('Choose how images are inserted into the editor.')
-			.addDropdown(dropdown => dropdown
-				.addOption('Procedual', 'Procedual (inserts images one by one)')
-				.addOption('Batch', 'Batch (inserts all images at once)')
-				.setValue(this.plugin.settings.insertionMethod)
-				.onChange(async (value) => {
-					this.plugin.settings.insertionMethod = value;
-					await this.plugin.saveSettings();
-				}));
-
-		// Empty Line setting
-		new Setting(containerEl)
-			.setName('Image separator')
-			.setDesc('Choose what to insert after each image.')
-			.addDropdown(dropdown => dropdown
-				.addOption('0', 'None')
-				.addOption('1', 'Empty line')
-				.addOption('2', 'Separator line')
-				.addOption('3', 'Empty line + separator line')
-				.setValue(this.plugin.settings.afterImage.toString())
-				.onChange(async (value) => {
-					this.plugin.settings.afterImage = parseInt(value, 10);
-					await this.plugin.saveSettings();
-				}));
-
-		// Enable Headers setting
-		new Setting(containerEl)
-			.setName('Insert headers (BETA)')
-			.setDesc('Finds headers in images and inserts them above the image. This is a beta feature and may not work as expected.')
-			.addToggle(toggle => toggle
-				.setValue(this.plugin.settings.enableHeaders)
-				.onChange(async (value) => {
-					this.plugin.settings.enableHeaders = value;
-					await this.plugin.saveSettings();
-					this.display(); // Refresh the settings page to show/hide the header size setting
-				}));
-
-		// Header advanced settings
-		if (this.plugin.settings.enableHeaders) {
-			// Remove Header Duplicates setting
-			new Setting(containerEl)
-			.setName('Remove header duplicates')
-			.setDesc('Removes duplicate headers from the image. This is useful if the same header appears on multiple pages.')
-			.addToggle(toggle => toggle
-				.setValue(this.plugin.settings.removeHeaderDuplicates)
-				.onChange(async (value) => {
-					this.plugin.settings.removeHeaderDuplicates = value;
-					await this.plugin.saveSettings();
-					this.display();
-				}));
-
-			// Header Size setting
-			new Setting(containerEl)
-				.setName('Header size')
-				.setDesc('The size of the header to be inserted above the image.')
-				.addDropdown(dropdown => dropdown
-					.addOption('#', 'h1')
-					.addOption('##', 'h2')
-					.addOption('###', 'h3')
-					.addOption('####', 'h4')
-					.addOption('#####', 'h5')
-					.setValue(this.plugin.settings.headerSize)
-					.onChange(async (value) => {
-						this.plugin.settings.headerSize = value;
-						await this.plugin.saveSettings();
-					})
-				);
-
-			// Header Extraction Sensitivity setting
-			new Setting(containerEl)
-				.setName('Header extraction sensitivity')
-				.setDesc('The sensitivity of the header extraction algorithm. Increase this value if headers are not being detected. Lower this value if non-headers are being detected as headers. The default is 1.2.')
-				.addSlider(slider => {
-					slider
-						.setLimits(0, 2, 0.1)
-						.setValue(this.plugin.settings.headerExtractionSensitive)
-						.setDynamicTooltip()
-						.onChange(async (value) => {
-							this.plugin.settings.headerExtractionSensitive = value;
-							await this.plugin.saveSettings();
-						});
-				});
-		}
-	}
-}
